--- app/src/main/java/com/stacktivity/yandeximagesearchengine/base/BaseViewModel.kt
@@ -1,9 +1,0 @@
-package com.stacktivity.yandeximagesearchengine.base
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-open class BaseViewModel : ViewModel() {
-    val empty = MutableLiveData<Boolean>().apply { value = true }
-    val dataLoading = MutableLiveData<Boolean>().apply { value = false }
-}
